feat(refactor): search button component
